Update: extract showToastOnUiThread to utils newly created utils module that contains all module common utils functions.
