+++ icons/build.gradle.kts
@@ -40,4 +40,9 @@ dependencies {
     val materialIconsVersion = "1.6.6"
     implementation("androidx.compose.material:material-icons-core:$materialIconsVersion")
     implementation("androidx.compose.material:material-icons-extended:$materialIconsVersion")
+
+    /**
+     * Utils
+     */
+    implementation(project(":utils"))
 }
