+++ settings.gradle.kts
@@ -20,3 +20,4 @@ include(":playback")
 include(":database")
 include(":internet")
 include(":icons")
+include(":utils")
