+++ utils/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-    namespace = "io.github.antoinepirlot.satunes.internet"
+    namespace = "io.github.antoinepirlot.utils"
     compileSdk = 34
 
     defaultConfig {
@@ -34,26 +34,10 @@ android {
 
 dependencies {
 
-    implementation("androidx.core:core-ktx:1.12.0")
+    implementation("androidx.core:core-ktx:1.13.0")
     implementation("androidx.appcompat:appcompat:1.6.1")
     implementation("com.google.android.material:material:1.11.0")
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-
-    /**
-     * Mutable State
-     */
-    implementation(platform("androidx.compose:compose-bom:2024.04.00"))
-    implementation("androidx.compose.material3:material3")
-
-    /**
-     * OkHttp3
-     */
-    implementation("com.squareup.okhttp3:okhttp:4.9.3")
-
-    /**
-     * Utils
-     */
-    implementation(project(":utils"))
 }
